Move open bucket hashtable into its own source file.
